Server refactoring
