+++ platform/dolphin-platform-server-spring/src/main/java/com/canoo/dolphin/server/spring/DolphinPlatformSpringBootstrap.java
@@ -17,7 +17,7 @@ package com.canoo.dolphin.server.spring;
 
 import com.canoo.dolphin.server.impl.ConfigurationFileLoader;
 import com.canoo.dolphin.server.config.DolphinPlatformConfiguration;
-import com.canoo.dolphin.server.servlet.DolphinPlatformBootstrap;
+import com.canoo.dolphin.server.bootstrap.DolphinPlatformBootstrap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.context.embedded.ServletContextInitializer;
