+++ platform/dolphin-platform-server/src/main/java/com/canoo/dolphin/server/context/DolphinContextFactory.java
@@ -1,0 +1,12 @@
+package com.canoo.dolphin.server.context;
+
+import javax.servlet.http.HttpSession;
+
+/**
+ * Created by hendrikebbers on 31.05.16.
+ */
+public interface DolphinContextFactory {
+
+    DolphinContext create(HttpSession httpSession, DolphinSessionListenerProvider dolphinSessionListenerProvider);
+
+}
