--- platform/dolphin-platform-server/src/main/java/com/canoo/dolphin/server/servlet/DolphinAccessException.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2015-2016 Canoo Engineering AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.canoo.dolphin.server.servlet;
-
-/**
- * Created by hendrikebbers on 02.05.16.
- */
-public class DolphinAccessException extends RuntimeException {
-
-    private static final long serialVersionUID = -2257268594556887415L;
-
-    public DolphinAccessException() {
-    }
-
-    public DolphinAccessException(String message) {
-        super(message);
-    }
-
-    public DolphinAccessException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public DolphinAccessException(Throwable cause) {
-        super(cause);
-    }
-}
